+++ Tests.Integration.Presentation.Web/Tools/BaseTest.cs
@@ -29,12 +29,15 @@ namespace Tests.Integration.Presentation.Web.Tools
     public class BaseTest : WithAutoFixture
     {
         private string _token;
-        public readonly Guid DefaultOrgUuid = DatabaseAccess.GetEntityUuid<Organization>(TestEnvironment.DefaultOrganizationId);
+
+        public readonly Guid DefaultOrgUuid =
+            DatabaseAccess.GetEntityUuid<Organization>(TestEnvironment.DefaultOrganizationId);
 
         public readonly Guid SecondOrgUuid =
             DatabaseAccess.GetEntityUuid<Organization>(TestEnvironment.SecondOrganizationId);
 
-        public async Task<ShallowOrganizationResponseDTO> CreateOrganizationAsync(string name = null, string cvr = null, OrganizationType type = OrganizationType.Municipality)
+        public async Task<ShallowOrganizationResponseDTO> CreateOrganizationAsync(string name = null, string cvr = null,
+            OrganizationType type = OrganizationType.Municipality)
         {
             var defaultRequest = new OrganizationCreateRequestDTO
             {
@@ -66,7 +69,8 @@ namespace Tests.Integration.Presentation.Web.Tools
             });
         }
 
-        public async Task<ItSystemResponseDTO> CreateItSystemAsync(Guid orgGuid, string name = null, RegistrationScopeChoice scope = RegistrationScopeChoice.Global, Guid? rightsHolderUuid = null)
+        public async Task<ItSystemResponseDTO> CreateItSystemAsync(Guid orgGuid, string name = null,
+            RegistrationScopeChoice scope = RegistrationScopeChoice.Global, Guid? rightsHolderUuid = null)
         {
             var request = new CreateItSystemRequestDTO
             {
@@ -96,7 +100,8 @@ namespace Tests.Integration.Presentation.Web.Tools
             }
         }
 
-        public async Task<OrganizationUnitResponseDTO> CreateOrganizationUnitAsync(Guid organizationUuid, string name = null, Guid? parentUnitUuid = null)
+        public async Task<OrganizationUnitResponseDTO> CreateOrganizationUnitAsync(Guid organizationUuid,
+            string name = null, Guid? parentUnitUuid = null)
         {
             var rootUnit = await OrganizationUnitV2Helper.GetRootUnit(organizationUuid);
             var request = new CreateOrganizationUnitRequestDTO
@@ -107,7 +112,8 @@ namespace Tests.Integration.Presentation.Web.Tools
             return await OrganizationUnitV2Helper.CreateUnitAsync(organizationUuid, request);
         }
 
-        public async Task<ItInterfaceResponseDTO> CreateItInterfaceAsync(Guid organizationUuid, string name = null, RegistrationScopeChoice scope = RegistrationScopeChoice.Global, string interfaceId = null)
+        public async Task<ItInterfaceResponseDTO> CreateItInterfaceAsync(Guid organizationUuid, string name = null,
+            RegistrationScopeChoice scope = RegistrationScopeChoice.Global, string interfaceId = null)
         {
             var request = new CreateItInterfaceRequestDTO
             {
@@ -166,6 +172,7 @@ namespace Tests.Integration.Presentation.Web.Tools
         public int GetOrgId(Guid orgUuid)
         {
             return DatabaseAccess.GetEntityId<Organization>(orgUuid);
-         
+
+        }
     }
 }
